requirements.txt
torch
torchvision
torchaudio
numpy
scipy
pandas
pillow
opencv-python
matplotlib
tqdm
scikit-image
scikit-learn
pyyaml
easydict
einops
timm
albumentations
diffusers
transformers
accelerate
safetensors
lpips
torch-fidelity
modelscope